Design: 검색창 구현
