+++ src/pages/index.tsx
@@ -8,3 +8,4 @@ export { default as ReservationPage } from './ReservationPage';
 export { default as SearchPage } from './SearchPage';
 export { default as UserPage } from './UserPage';
 export { default as NotFoundPage } from './NotFoundPage';
+export { default as Comments } from './Comments';
