+++ src/routes/Router.tsx
@@ -11,6 +11,7 @@ import {
   SearchPage,
   UserPage,
   NotFoundPage,
+  Comments,
 } from '../pages';
 
 const Router = () => {
@@ -25,6 +26,7 @@ const Router = () => {
       <Route path="/reservation" exact component={ReservationPage} />
       <Route path="/user" exact component={UserPage} />
       <Route path="/author" exact component={AuthorPage} />
+      <Route path="/feed/comments/:id" exact component={Comments} />
       <Route path="*" component={NotFoundPage} />
     </Switch>
   );
